# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Four Delivery API",
        "description": "API para o sistema de delivery Four Delivery",
        "contact": {
            "email": "suporte@fourdelivery.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/admin-v1",
            "description": "Admin API V1"
        },
        {
            "url": "/api/app-v1",
            "description": "App API V1"
        }
    ],
    "paths": {
...
        "/customers": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Listar clientes da empresa (paginado)",
                "operationId": "1171e97ffd82e321d87ada9c254f0324",
                "responses": {
                    "200": {
                        "description": "Lista paginada de clientes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Cadastrar novo cliente",
                "operationId": "8601a26c46c986ca44d13b83a75103dd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "phone",
                                    "address",
                                    "neighborhood",
                                    "city",
                                    "state",
                                    "zip_code"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "neighborhood": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "state": {
                                        "type": "string"
                                    },
                                    "zip_code": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente cadastrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/customers-all": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Listar todos os clientes (com ou sem paginação dependendo do cargo)",
                "operationId": "d811cc9d22850449c3037521134a9ce7",
                "responses": {
                    "200": {
                        "description": "Lista de clientes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/customers/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Mostrar detalhes de um cliente",
                "operationId": "acd69b97ac8773db061a3dffd0ebad79",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do cliente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Atualizar um cliente",
                "operationId": "bd621281c1cc1d28fd326dcf8928565f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente atualizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Excluir um cliente",
                "operationId": "d44a167f3af7c4b7fdbb753244006a1a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Cliente excluído"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/customers/search": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Pesquisar clientes por nome ou telefone",
                "operationId": "00e2ec556ee4193c41c1ab830d5c0c32",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados da pesquisa"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}